requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.117.1
Flask==3.1.2
flask-cors==6.0.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
npm==0.1.1
optional-django==0.1.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3